remove winrt list hack
PyWinRT v3.1 fixed the crash that was being worked around by by casting
the get_history_with_id() result to a list.

Update the minimal version of pywinrt to 3.1 and remove the list hack.
Also drop upper bound for packages versions as it can cause issues
for people consuming this library.